Add dummy field to Pawns::Entry to measure slowdown
The Pawns::Entry struct is exactly 128 bytes now. This happens to be
exactly the size of cache line on common CPUs. I suspect that just by
adding a new field to this structure we cause a slowdown, leading to an
almost 0 chance of passing for any patch that needs to add a new field
to Pawns::Entry. Let me try to measure the ELO cost of such addition.

No funtional change.