Invert counter value
As @locutus2 corrected me, we care mainly for cases where the counter is
nearing reset, so let's work with an inverted counter value instead.

High rule50 values are rare, so we can also raise the max shift all the
way up to 6 without even affecting the default bench.

Bench: 4248710 (same as master)